requirements.txt
asgiref==3.7.2
Babel==2.12.1
build==0.10.0
Django==4.2.2
django-ckeditor==6.5.1
django-environ==0.10.0
django-js-asset==2.0.0
django-money==3.1.0
mysqlclient==2.2.0
packaging==23.1
#Pillow==9.5.0
psycopg2-binary==2.9.9
py-moneyed==2.0
pyproject_hooks==1.0.0
sqlparse==0.4.4
typing_extensions==4.6.3
